+++ packages/ui5-tooling-modules/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.6](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.7.5...ui5-tooling-modules@0.7.6) (2023-01-21)
+
+### Bug Fixes
+
+- **ui5-tooling-modules:** hide warnings ([04f7efb](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/04f7efb241a7ede65963fcd4b8a23d66f3c7f83c))
+
+## [0.7.5](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.7.4...ui5-tooling-modules@0.7.5) (2023-01-21)
+
+### Bug Fixes
+
+- **ui5-tooling-modules:** rollback to old node-resolve ([e1b418f](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/e1b418fc207919054af40e1ce1afba545847a42e))
+
+## [0.7.4](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.7.3...ui5-tooling-modules@0.7.4) (2023-01-21)
+
+### Bug Fixes
+
+- **ui5-tooling-modules:** ignore relative paths ([#676](https://github.com/ui5-community/ui5-ecosystem-showcase/issues/676)) ([259aac7](https://github.com/ui5-community/ui5-ecosystem-showcase/commit/259aac754233925bed6bdbd1f79a13757786b1c6))
+
 ## [0.7.3](https://github.com/ui5-community/ui5-ecosystem-showcase/compare/ui5-tooling-modules@0.7.2...ui5-tooling-modules@0.7.3) (2023-01-08)
 
 ### Bug Fixes